src/requirements.txt
aiogram==2.12.1
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==20.3.0
Babel==2.9.0
certifi==2020.12.5
chardet==4.0.0
idna==3.1
multidict==5.1.0
python-dotenv==0.16.0
pytz==2021.1
typing-extensions==3.7.4.3
yarl==1.6.3
